add a TMPL_CONST semantic
TMPL_CONST defines a constant in template with symbol and its value as
follows:

<!-- TMPL_CONST foo "bar baz" -->

The above would define a symbol `foo` with value of "bar baz", which
is added to template printer values, and can be used as variable in
subsequence.

I add this to enable me to define constant page title, goback links,
etc. in template file, rather than in lisp code. I think they are
specific only to html page rendering, but no related to business
logic, would be preferably defined in html template files.